+++ src/shared/const/comment.ts
@@ -1,3 +1,5 @@
+export const MAX_COMMENTS_COUNT = 50;
+
 export enum TextLength {
   Min = 1,
   Max = 15,
