+++ src/shared/libs/rest/errors/validation.error.ts
@@ -1,6 +1,6 @@
 import { StatusCodes } from 'http-status-codes';
-import { HttpError } from './http-error.js';
-import { ValidationErrorField } from '../types/validation-error-field.type.js';
+import { HttpError } from './index.js';
+import { ValidationErrorField } from '../index.js';
 
 export class ValidationError extends HttpError {
   public details: ValidationErrorField[] = [];
