+++ src/shared/libs/rest/exception-filter/app.exception-filter.ts
@@ -1,11 +1,10 @@
 import { inject, injectable } from 'inversify';
 import { StatusCodes } from 'http-status-codes';
 import { NextFunction, Request, Response } from 'express';
-import { ExceptionFilter } from './exception-filter.interface.js';
+import { ExceptionFilter, ApplicationError } from '../index.js';
 import { Logger } from '../../logger/index.js';
 import { Component } from '../../../types/index.js';
 import { createErrorObject } from '../../../helpers/index.js';
-import { ApplicationError } from '../types/application-error.enum.js';
 
 @injectable()
 export class AppExceptionFilter implements ExceptionFilter {
