+++ src/shared/libs/rest/exception-filter/http-error.exception-filter.ts
@@ -6,7 +6,7 @@ import { Component } from '../../../types/index.js';
 import { Logger } from '../../logger/index.js';
 import { HttpError } from '../errors/index.js';
 import { createErrorObject } from '../../../helpers/index.js';
-import { ApplicationError } from '../types/application-error.enum.js';
+import { ApplicationError } from '../index.js';
 
 @injectable()
 export class HttpErrorExceptionFilter implements ExceptionFilter {
