+++ src/shared/libs/rest/types/route.interface.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
 import { HttpMethod } from './http-method.enum.js';
-import { Middleware } from '../middleware/middleware.interface.js';
+import { Middleware } from '../index.js';
 
 export interface Route {
   path: string;
