+++ src/shared/modules/auth/auth.container.ts
@@ -1,9 +1,8 @@
 import { Container } from 'inversify';
-import { AuthService } from './auth-service.interface.js';
+import { AuthService, DefaultAuthService } from './index.js';
 import { Component } from '../../types/index.js';
-import { DefaultAuthService } from './default-auth.service.js';
 import { ExceptionFilter } from '../../libs/rest/index.js';
-import { AuthExceptionFilter } from './errors/auth.exception-filter.js';
+import { AuthExceptionFilter } from './errors/index.js';
 
 export function createAuthContainer() {
   const authContainer = new Container();
