+++ src/shared/modules/auth/errors/index.ts
@@ -1,3 +1,4 @@
 export * from './base-user.exception.js';
 export * from './user-not-found.exception.js';
 export * from './user-password-incorrect.exception.js';
+export * from './auth.exception-filter.js';
