+++ src/shared/modules/comment/comment-service.interface.ts
@@ -1,6 +1,6 @@
 import { CreateCommentDto } from './dto/create-comment.dto.js';
 import { DocumentType } from '@typegoose/typegoose';
-import { CommentEntity } from './comment.entity.js';
+import { CommentEntity } from './index.js';
 
 export interface CommentService {
   create(dto: CreateCommentDto, offerId: string, userId: string): Promise<DocumentType<CommentEntity>>;
