--- src/shared/modules/comment/comment.constant.ts
@@ -1,1 +1,0 @@
-export const MAX_COMMENTS_COUNT = 50;
