+++ src/shared/modules/comment/comment.container.ts
@@ -1,9 +1,7 @@
 import { Container } from 'inversify';
 import { types } from '@typegoose/typegoose';
-import { CommentService } from './comment-service.interface.js';
+import { CommentService, DefaultCommentService, CommentEntity, CommentModel } from './index.js';
 import { Component } from '../../types/index.js';
-import { CommentEntity, CommentModel } from './comment.entity.js';
-import { DefaultCommentService } from './default-comment.service.js';
 import CommentController from './comment.controller.js';
 import { Controller } from '../../libs/rest/index.js';
