+++ src/shared/modules/comment/comment.controller.ts
@@ -4,12 +4,10 @@ import { StatusCodes } from 'http-status-codes';
 import { BaseController, HttpError, HttpMethod, ValidateDtoMiddleware, PrivateRouteMiddleware, } from '../../libs/rest/index.js';
 import { Component } from '../../types/index.js';
 import { Logger } from '../../libs/logger/index.js';
-import { CommentService } from './comment-service.interface.js';
 import { OfferService } from '../offer/index.js';
 import { fillDTO } from '../../helpers/index.js';
 import { CommentRdo } from './rdo/comment.rdo.js';
-import { CreateCommentRequest } from './types/create-comment-request.type.js';
-import { CreateCommentDto } from './dto/create-comment.dto.js';
+import { CommentService, CreateCommentDto, CreateCommentRequest } from './index.js';
 
 @injectable()
 export default class CommentController extends BaseController {
