+++ src/shared/modules/comment/index.ts
@@ -4,3 +4,4 @@ export * from './comment.entity.js';
 export * from './default-comment.service.js';
 export * from './dto/create-comment.dto.js';
 export * from './rdo/comment.rdo.js';
+export * from './types/create-comment-request.type.js';
