+++ src/shared/modules/comment/rdo/comment.rdo.ts
@@ -1,5 +1,5 @@
 import { Expose, Type } from 'class-transformer';
-import { UserRdo } from '../../user/rdo/user.rdo.js';
+import { UserRdo } from '../../user/index.js';
 
 export class CommentRdo {
   @Expose()
