--- src/shared/modules/offer/offer.constant.ts
@@ -1,35 +1,0 @@
-export const DEFAULT_OFFERS_COUNT = 60;
-
-export const MAX_PREMIUM_OFFERS_COUNT = 3;
-
-export const Cities = {
-  Paris: {
-    latitude: 48.85661,
-    longitude:  2.351499
-  },
-
-  Cologne: {
-    latitude: 50.938361,
-    longitude: 6.959974
-  },
-
-  Brussels: {
-    latitude: 50.846557,
-    longitude: 4.351697
-  },
-
-  Amsterdam: {
-    latitude: 52.370216,
-    longitude: 4.895168
-  },
-
-  Hamburg: {
-    latitude: 53.550341,
-    longitude: 10.000654
-  },
-
-  Dusseldorf: {
-    latitude: 53.550341,
-    longitude: 10.000654
-  },
-} as const;
