+++ src/shared/modules/offer/offer.container.ts
@@ -1,8 +1,6 @@
 import { Container } from 'inversify';
-import { OfferService } from './offer-service.interface.js';
 import { Component } from '../../types/index.js';
-import { DefaultOfferService } from './default-offer.service.js';
-import { OfferEntity, OfferModel } from './offer.entity.js';
+import { OfferService, DefaultOfferService, OfferEntity, OfferModel } from './index.js';
 import { types } from '@typegoose/typegoose';
 import OfferController from './offer.controller.js';
 import { Controller } from '../../libs/rest/index.js';
