+++ src/shared/modules/offer/rdo/fulloffer.rdo.ts
@@ -1,5 +1,5 @@
 import { Expose, Type } from 'class-transformer';
-import { UserRdo } from '../../user/rdo/user.rdo.js';
+import { UserRdo } from '../../user/index.js';
 import { City, Goods, Property } from '../../../types/index.js';
 
 export class FullOfferRdo {
