+++ src/shared/modules/offer/type/create-offer-request.type.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
 import { RequestBody, RequestParams } from '../../../libs/rest/index.js';
-import { CreateOfferDto } from '../dto/create-offer.dto.js';
+import { CreateOfferDto } from '../index.js';
 
 export type CreateOfferRequest = Request<RequestParams, RequestBody, CreateOfferDto>;
