+++ src/shared/modules/user/create-user-request.type.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
 import { RequestBody, RequestParams } from '../../libs/rest/index.js';
-import { CreateUserDto } from './dto/create-user.dto.js';
+import { CreateUserDto } from './index.js';
 
 export type CreateUserRequest = Request<RequestParams, RequestBody, CreateUserDto>;
