+++ src/shared/modules/user/index.ts
@@ -1,8 +1,14 @@
-export * from './user.entity.js';
 export * from './dto/create-user.dto.js';
+export * from './dto/login-user.dto.js';
+export * from './dto/update-user.dto.js';
+export * from './rdo/user.rdo.js';
+export * from './rdo/login-user.rdo.js';
+export * from './rdo/logged-user.rdo.js';
+export * from './rdo/upload-user-avatar.rdo.js';
+export * from './user.entity.js';
 export * from './default-user.service.js';
 export * from './user.container.js';
 export * from './user.controller.js';
-export * from './dto/login-user.dto.js';
 export * from './user-service.interface.js';
-export * from './user.constant.js';
+export * from './create-user-request.type.js';
+export * from './login-user-request.type.js';
