+++ src/shared/modules/user/login-user-request.type.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
 import { RequestBody, RequestParams } from '../../libs/rest/index.js';
-import { LoginUserDto } from './dto/login-user.dto.js';
+import { LoginUserDto } from './index.js';
 
 export type LoginUserRequest = Request<RequestParams, RequestBody, LoginUserDto>;
