+++ src/shared/modules/user/rdo/user.rdo.ts
@@ -1,5 +1,5 @@
 import { Expose } from 'class-transformer';
-import { UserType } from '../../../types/user.type.js';
+import { UserType } from '../../../types/index.js';
 
 export class UserRdo {
   @Expose()
