--- src/shared/modules/user/user.constant.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_AVATAR_FILE_NAME = 'default-avatar.jpg';
