+++ src/shared/types/comment.type.ts
@@ -1,4 +1,4 @@
-import { User } from './user.type.js';
+import { User } from './index.js';
 
 export type Comment = {
   text: string;
